Removed the SCION bandwidth-test client as SEED service.
