+++ seedemu/services/__init__.py
@@ -9,3 +9,4 @@ from .CymruIpOrigin import CymruIpOriginService, CymruIpOriginServer
 from .ReverseDomainNameService import ReverseDomainNameService, ReverseDomainNameServer
 from .BgpLookingGlassService import BgpLookingGlassServer, BgpLookingGlassService
 from .DHCPService import DHCPServer, DHCPService
+from .ScionBwtestServerService import ScionBwtestServerService
